Move signature module to config file
